appengine/findit/requirements.txt
# This file is generated from the .vpython3 spec file.
# Use `make generate_requirements_txt` to update.
appengine-python-standard==1.1.1
attrs==21.4.0
beautifulsoup4==4.9.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.4
Click==7.0
Flask==1.0.2
frozendict==2.0.6
google-api-core==1.31.5
google-api-python-client==2.2.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.2
google-cloud-logging==3.0.0
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
httplib2==0.19.1
idna==2.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==4.0.3
oauth2client==4.1.3
packaging==23.0
parameterized==0.8.1
Pillow==8.3.1
pluggy==0.13.1
proto-plus==1.20.3
protobuf==3.20.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
pytest-rerunfailures==11.1.2
pytz==2021.1
requests==2.31.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.15.0
soupsieve==1.9.5
toml==0.10.1
uritemplate==3.0.0
urllib3==1.26.4
waitress==1.4.3
WebOb==1.8.6
WebTest==2.0.35
Werkzeug==1.0.1